set up boilerplate for Redux Store and Combined Reducers
