+++ client/redux/actions/constants/actionTypes.js
@@ -1,11 +1,8 @@
 // Action Types - What we're updating in State
 
-
-//overall reducer
+// userReducer
 export const IS_LOGGED_IN = 'IS_LOGGED_IN';
 export const IS_LOGGED_OUT = 'IS_LOGGED_OUT';
-
-// userReducer
 export const USER_INFO = 'USER_INFO';
 export const ADD_HOUSEHOLD_TO_USER = 'ADD_HOUSE_TO_USER';
 export const UPDATE_USER_ITEMS = 'UPDATE_USER_ITEMS'
